monitoring/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
scikit-learn
matplotlib
python-dotenv
seaborn